Parallel Implementation Done. Not working yet

+++ src/main/java/hyflow/common/RequestId.java
@@ -63,9 +63,9 @@ public class RequestId implements Serializable, Comparable<RequestId> {
 
     public int compareTo(RequestId requestId) {
         if (clientId != requestId.clientId)
-            return (int) (clientId - requestId.clientId);
+            return clientId - requestId.clientId;
         else
-            return (int) (seqNumber - requestId.seqNumber);
+            return seqNumber - requestId.seqNumber;
     }
 
     public boolean equals(Object obj) {
